Fix `ALTERNATIVE_QUEUE`.

+++ src/httplib_start.c
@@ -172,10 +172,10 @@ struct httplib_context *httplib_start( const struct httplib_callbacks *callbacks
 
 #if defined(ALTERNATIVE_QUEUE)
 
-		ctx->client_wait_events = httplib_calloc( sizeof(ctx->client_wait_events[0]), ctx->num_threads );
+		ctx->client_wait_events = httplib_calloc( sizeof(ctx->client_wait_events[0]), (size_t)ctx->num_threads );
 		if ( ctx->client_wait_events == NULL ) return XX_httplib_abort_start( ctx, "Not enough memory for worker event array" );
 
-		ctx->client_socks = httplib_calloc( sizeof(ctx->client_socks[0]), ctx->num_threads );
+		ctx->client_socks = httplib_calloc( sizeof(ctx->client_socks[0]), (size_t)ctx->num_threads );
 		if ( ctx->client_socks == NULL ) return XX_httplib_abort_start( ctx, "Not enough memory for worker socket array" );
 
 		for (i=0; i<ctx->num_threads; i++) {
